Move playground functions to helper.js files
